# openfinance-consents-service/src/main/resources/openapi/consents-3.2.0.yml
openapi: 3.0.1
info:
  title: API Consents - Open Finance Brasil
  description: API de consentimentos para o Open Finance Brasil
  version: 3.2.0
  contact:
    name: Open Finance Brasil
    url: https://openfinance.org.br
    email: contato@openfinance.org.br
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

servers:
  - url: https://api.openfinance.org.br/open-banking/consents/v3
    description: Servidor de produção
  - url: https://api.sandbox.openfinance.org.br/open-banking/consents/v3
    description: Servidor de sandbox

paths:
  /consents:
    post:
      tags:
        - Consents
      summary: Criar novo consentimento
      operationId: createConsent
      description: Cria novo consentimento para compartilhamento de dados
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/xFapiAuthDate'
        - $ref: '#/components/parameters/xFapiCustomerIpAddress'
        - $ref: '#/components/parameters/xFapiInteractionId'
        - $ref: '#/components/parameters/xCustomerUserAgent'
        - $ref: '#/components/parameters/xJwsSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConsentRequest'
      responses:
        '201':
          description: Consentimento criado com sucesso
          headers:
            x-fapi-interaction-id:
              $ref: '#/components/headers/xFapiInteractionId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseConsentCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
      security:
        - OAuth2Security:
            - consents

  /consents/{consentId}:
    get:
      tags:
        - Consents
      summary: Obter detalhes do consentimento
      operationId: getConsent
      description: Obtém detalhes de um consentimento específico
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/xFapiAuthDate'
        - $ref: '#/components/parameters/xFapiCustomerIpAddress'
        - $ref: '#/components/parameters/xFapiInteractionId'
        - $ref: '#/components/parameters/xCustomerUserAgent'
        - $ref: '#/components/parameters/consentId'
      responses:
        '200':
          description: Consentimento obtido com sucesso
          headers:
            x-fapi-interaction-id:
              $ref: '#/components/headers/xFapiInteractionId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseConsent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
      security:
        - OAuth2Security:
            - consents

    delete:
      tags:
        - Consents
      summary: Revogar consentimento
      operationId: deleteConsent
      description: Revoga um consentimento específico
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/xFapiAuthDate'
        - $ref: '#/components/parameters/xFapiCustomerIpAddress'
        - $ref: '#/components/parameters/xFapiInteractionId'
        - $ref: '#/components/parameters/xCustomerUserAgent'
        - $ref: '#/components/parameters/consentId'
      responses:
        '204':
          description: Consentimento revogado com sucesso
          headers:
            x-fapi-interaction-id:
              $ref: '#/components/headers/xFapiInteractionId'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
      security:
        - OAuth2Security:
            - consents

  /consents/{consentId}/extends:
    post:
      tags:
        - Consents
      summary: Renovar consentimento
      operationId: extendConsent
      description: Renova um consentimento existente
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/xFapiAuthDate'
        - $ref: '#/components/parameters/xFapiCustomerIpAddress'
        - $ref: '#/components/parameters/xFapiInteractionId'
        - $ref: '#/components/parameters/xCustomerUserAgent'
        - $ref: '#/components/parameters/consentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtendConsentRequest'
      responses:
        '200':
          description: Consentimento renovado com sucesso
          headers:
            x-fapi-interaction-id:
              $ref: '#/components/headers/xFapiInteractionId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseConsentExtended'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
      security:
        - OAuth2Security:
            - consents

  /consents/{consentId}/extensions:
    get:
      tags:
        - Consents
      summary: Obter histórico de extensões
      operationId: getConsentExtensions
      description: Obtém histórico de extensões de um consentimento
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/xFapiAuthDate'
        - $ref: '#/components/parameters/xFapiCustomerIpAddress'
        - $ref: '#/components/parameters/xFapiInteractionId'
        - $ref: '#/components/parameters/xCustomerUserAgent'
        - $ref: '#/components/parameters/consentId'
      responses:
        '200':
          description: Histórico de extensões obtido com sucesso
          headers:
            x-fapi-interaction-id:
              $ref: '#/components/headers/xFapiInteractionId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseConsentExtensions'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
      security:
        - OAuth2Security:
            - consents

components:
  schemas:
    CreateConsentRequest:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/ConsentData'

    ConsentData:
      type: object
      required:
        - loggedUser
        - businessEntity
        - permissions
        - expirationDateTime
      properties:
        loggedUser:
          $ref: '#/components/schemas/LoggedUser'
        businessEntity:
          $ref: '#/components/schemas/BusinessEntity'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/EnumConsentPermissions'
          minItems: 1
        expirationDateTime:
          type: string
          format: date-time
          description: Data e hora de expiração do consentimento
        transactionFromDateTime:
          type: string
          format: date-time
          description: Data e hora de início para consulta de transações
        transactionToDateTime:
          type: string
          format: date-time
          description: Data e hora de fim para consulta de transações

    LoggedUser:
      type: object
      required:
        - document
      properties:
        document:
          $ref: '#/components/schemas/Document'

    BusinessEntity:
      type: object
      properties:
        document:
          $ref: '#/components/schemas/Document'

    Document:
      type: object
      required:
        - identification
        - rel
      properties:
        identification:
          type: string
          description: CPF ou CNPJ do usuário
          pattern: '^\d{11}$|^\d{14}$'
        rel:
          type: string
          description: Tipo de documento
          enum:
            - CPF
            - CNPJ

    EnumConsentPermissions:
      type: string
      enum:
        - ACCOUNTS_READ
        - ACCOUNTS_BALANCES_READ
        - ACCOUNTS_TRANSACTIONS_READ
        - ACCOUNTS_OVERDRAFT_LIMITS_READ
        - CREDIT_CARDS_ACCOUNTS_READ
        - CREDIT_CARDS_ACCOUNTS_BILLS_READ
        - CREDIT_CARDS_ACCOUNTS_BILLS_TRANSACTIONS_READ
        - CREDIT_CARDS_ACCOUNTS_LIMITS_READ
        - CREDIT_CARDS_ACCOUNTS_TRANSACTIONS_READ
        - CUSTOMERS_PERSONAL_IDENTIFICATIONS_READ
        - CUSTOMERS_PERSONAL_ADITTIONALINFO_READ
        - CUSTOMERS_BUSINESS_IDENTIFICATIONS_READ
        - CUSTOMERS_BUSINESS_ADITTIONALINFO_READ
        - FINANCINGS_READ
        - FINANCINGS_SCHEDULED_INSTALMENTS_READ
        - FINANCINGS_PAYMENTS_READ
        - FINANCINGS_WARRANTIES_READ
        - INVOICE_FINANCINGS_READ
        - INVOICE_FINANCINGS_SCHEDULED_INSTALMENTS_READ
        - INVOICE_FINANCINGS_PAYMENTS_READ
        - INVOICE_FINANCINGS_WARRANTIES_READ
        - LOANS_READ
        - LOANS_SCHEDULED_INSTALMENTS_READ
        - LOANS_PAYMENTS_READ
        - LOANS_WARRANTIES_READ
        - UNARRANGED_ACCOUNTS_OVERDRAFT_READ
        - UNARRANGED_ACCOUNTS_OVERDRAFT_SCHEDULED_INSTALMENTS_READ
        - UNARRANGED_ACCOUNTS_OVERDRAFT_PAYMENTS_READ
        - UNARRANGED_ACCOUNTS_OVERDRAFT_WARRANTIES_READ
        - BANK_FIXED_INCOMES_READ
        - CREDIT_FIXED_INCOMES_READ
        - FUNDS_READ
        - VARIABLE_INCOMES_READ
        - TREASURE_TITLES_READ
        - EXCHANGES_READ

    ResponseConsentCreated:
      type: object
      required:
        - data
        - links
        - meta
      properties:
        data:
          $ref: '#/components/schemas/ConsentCreatedData'
        links:
          $ref: '#/components/schemas/Links'
        meta:
          $ref: '#/components/schemas/Meta'

    ConsentCreatedData:
      type: object
      required:
        - consentId
        - status
        - statusUpdateDateTime
        - creationDateTime
        - expirationDateTime
        - permissions
      properties:
        consentId:
          type: string
          description: Identificador único do consentimento
          maxLength: 256
        status:
          $ref: '#/components/schemas/EnumConsentStatus'
        statusUpdateDateTime:
          type: string
          format: date-time
          description: Data e hora da última atualização do status
        creationDateTime:
          type: string
          format: date-time
          description: Data e hora de criação do consentimento
        expirationDateTime:
          type: string
          format: date-time
          description: Data e hora de expiração do consentimento
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/EnumConsentPermissions'
          minItems: 1

    ResponseConsent:
      type: object
      required:
        - data
        - links
        - meta
      properties:
        data:
          $ref: '#/components/schemas/ConsentFullData'
        links:
          $ref: '#/components/schemas/Links'
        meta:
          $ref: '#/components/schemas/Meta'

    ConsentFullData:
      type: object
      required:
        - consentId
        - status
        - statusUpdateDateTime
        - creationDateTime
        - expirationDateTime
        - permissions
      properties:
        consentId:
          type: string
          description: Identificador único do consentimento
          maxLength: 256
        status:
          $ref: '#/components/schemas/EnumConsentStatus'
        statusUpdateDateTime:
          type: string
          format: date-time
          description: Data e hora da última atualização do status
        creationDateTime:
          type: string
          format: date-time
          description: Data e hora de criação do consentimento
        expirationDateTime:
          type: string
          format: date-time
          description: Data e hora de expiração do consentimento
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/EnumConsentPermissions'
          minItems: 1
        loggedUser:
          $ref: '#/components/schemas/LoggedUser'
        businessEntity:
          $ref: '#/components/schemas/BusinessEntity'
        transactionFromDateTime:
          type: string
          format: date-time
          description: Data e hora de início para consulta de transações
        transactionToDateTime:
          type: string
          format: date-time
          description: Data e hora de fim para consulta de transações

    ExtendConsentRequest:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/ConsentExtendData'

    ConsentExtendData:
      type: object
      required:
        - expirationDateTime
      properties:
        expirationDateTime:
          type: string
          format: date-time
          description: Nova data e hora de expiração do consentimento

    ResponseConsentExtended:
      type: object
      required:
        - data
        - links
        - meta
      properties:
        data:
          $ref: '#/components/schemas/ConsentExtendedData'
        links:
          $ref: '#/components/schemas/Links'
        meta:
          $ref: '#/components/schemas/Meta'

    ConsentExtendedData:
      type: object
      required:
        - consentId
        - expirationDateTime
        - statusUpdateDateTime
      properties:
        consentId:
          type: string
          description: Identificador único do consentimento
          maxLength: 256
        expirationDateTime:
          type: string
          format: date-time
          description: Nova data e hora de expiração do consentimento
        statusUpdateDateTime:
          type: string
          format: date-time
          description: Data e hora da última atualização do status

    ResponseConsentExtensions:
      type: object
      required:
        - data
        - links
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ConsentExtensionData'
        links:
          $ref: '#/components/schemas/Links'
        meta:
          $ref: '#/components/schemas/Meta'

    ConsentExtensionData:
      type: object
      required:
        - requestDateTime
        - expirationDateTime
        - previousExpirationDateTime
      properties:
        requestDateTime:
          type: string
          format: date-time
          description: Data e hora da solicitação de extensão
        expirationDateTime:
          type: string
          format: date-time
          description: Nova data e hora de expiração
        previousExpirationDateTime:
          type: string
          format: date-time
          description: Data e hora de expiração anterior

    EnumConsentStatus:
      type: string
      enum:
        - AWAITING_AUTHORISATION
        - AUTHORISED
        - REJECTED
        - REVOKED
        - EXPIRED
        - CONSUMED

    Links:
      type: object
      properties:
        self:
          type: string
          description: Link para a própria página
        first:
          type: string
          description: Link para a primeira página
        prev:
          type: string
          description: Link para a página anterior
        next:
          type: string
          description: Link para a próxima página
        last:
          type: string
          description: Link para a última página

    Meta:
      type: object
      required:
        - totalRecords
        - totalPages
      properties:
        totalRecords:
          type: integer
          description: Número total de registros
        totalPages:
          type: integer
          description: Número total de páginas
        requestDateTime:
          type: string
          format: date-time
          description: Data e hora da requisição

    Error:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - code
              - title
              - detail
            properties:
              code:
                type: string
                description: Código do erro
              title:
                type: string
                description: Título do erro
              detail:
                type: string
                description: Detalhes do erro
              meta:
                type: object
                description: Metadados do erro
        meta:
          $ref: '#/components/schemas/Meta'

  parameters:
    Authorization:
      name: Authorization
      in: header
      required: true
      description: Token de autorização
      schema:
        type: string
        pattern: '^Bearer\s'

    xFapiAuthDate:
      name: x-fapi-auth-date
      in: header
      required: true
      description: Data de autenticação
      schema:
        type: string
        format: date-time

    xFapiCustomerIpAddress:
      name: x-fapi-customer-ip-address
      in: header
      required: true
      description: Endereço IP do cliente
      schema:
        type: string

    xFapiInteractionId:
      name: x-fapi-interaction-id
      in: header
      required: true
      description: Identificador da interação
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]{8}-[a-zA-Z0-9]{4}-[a-zA-Z0-9]{4}-[a-zA-Z0-9]{4}-[a-zA-Z0-9]{12}$'

    xCustomerUserAgent:
      name: x-customer-user-agent
      in: header
      required: true
      description: User agent do cliente
      schema:
        type: string

    xJwsSignature:
      name: x-jws-signature
      in: header
      required: true
      description: Assinatura JWS da requisição
      schema:
        type: string

    consentId:
      name: consentId
      in: path
      required: true
      description: Identificador único do consentimento
      schema:
        type: string

  headers:
    xFapiInteractionId:
      description: Identificador da interação
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]{8}-[a-zA-Z0-9]{4}-[a-zA-Z0-9]{4}-[a-zA-Z0-9]{4}-[a-zA-Z0-9]{12}$'

  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Unauthorized:
      description: Token de autorização inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Forbidden:
      description: Acesso negado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    MethodNotAllowed:
      description: Método não permitido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    NotAcceptable:
      description: Não aceito
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    UnsupportedMediaType:
      description: Tipo de mídia não suportado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    UnprocessableEntity:
      description: Entidade não processável
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    TooManyRequests:
      description: Muitas requisições
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    InternalServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    GatewayTimeout:
      description: Timeout do gateway
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  securitySchemes:
    OAuth2Security:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.openfinance.org.br/oauth/token
          scopes:
            consents: Gerenciamento de consentimentos
        authorizationCode:
          authorizationUrl: https://auth.openfinance.org.br/oauth/authorize
          tokenUrl: https://auth.openfinance.org.br/oauth/token
          scopes:
            consents: Gerenciamento de consentimentos